move instance-initializer to addon directory
